refactor: to a seperate api package
