+++ apps/www/src/api/contract.ts
@@ -1,6 +1,6 @@
 import { initContract } from "@ts-rest/core";
 import { z } from "zod";
-import { cacheControlsQueryDtoSchema, errorSchemaRo } from "./schemas";
+import { cacheControlsQueryDtoSchema, errorSchemaRo } from "./rest/schemas";
 
 const c = initContract();
